コントローラー 修正

+++ app/controllers/items_controller.rb
@@ -35,8 +35,10 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
+    if item.user == current_user
+       item.destroy
     redirect_to root_path
+    end
   end
 
   private
